+++ packages/nextjs/components/moon/index.js
@@ -1,3 +1,1 @@
-export * from "./signup";
-export * from "./signin";
 export * from "./sign";
